requirements.txt
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.60.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pybaselines==1.2.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.13.1
six==1.17.0
tkinterdnd2-universal==1.7.3
tzdata==2025.2
zipp==3.23.0
